Title: 동일하게 분할, Time: 17ms, Memory: 0MB, Status: Runtime Error - Codetree

+++ 250722/동일하게 분할/equal-partition.cpp
@@ -3,11 +3,11 @@
 using namespace std;
 
 int n;
-int arr[101];
+long long arr[101];
 int dp[1001] = {0};
 int main() {
     cin >> n;
-    int total = 0;
+    long long total = 0;
     for (int i = 0; i < n; i++) {
         cin >> arr[i];
         total += arr[i];
@@ -17,8 +17,8 @@ int main() {
         return 0;
     }
     dp[0] = 1;
-    for (int i = 0; i < n; i++) {
-        for (int j = total; j >= arr[i]; j--) {
+    for (long long i = 0; i < n; i++) {
+        for (long long j = total; j >= arr[i]; j--) {
             if (dp[j - arr[i]]) {
                 dp[j]++;
             }
